requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-tastypie==0.14.3
kombu==5.1.0
Pillow==8.4.0
prompt-toolkit==3.0.21
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
